+++ src/common/hevc.h
@@ -434,7 +434,7 @@ protected:
 
   int64_t m_stream_default_duration, m_forced_default_duration, m_container_default_duration;
   int m_frame_number, m_num_skipped_frames;
-  bool m_first_keyframe_found, m_recovery_point_valid, m_b_frames_since_keyframe;
+  bool m_first_keyframe_found, m_recovery_point_valid, m_b_frames_since_keyframe, m_first_cleanup;
 
   bool m_par_found;
   int64_rational_c m_par;